// src/app/types/aboutDeveloper.ts
export interface aboutDeveloper{
    section:string;
    data:[]
}

export interface Developer {
    id: number;
    name: string;
    title: string;
    imageSrc: string;
    skills: string[];
    services: string[];
    contactNumber: string;
    email: string;
    youtubeLink: string;
    facebookLink: string;
    instagramLink: string;
}